Make going/being critical more noticeable
