+++ Content.Server/Mobs/Components/DeathgaspComponent.cs
@@ -6,7 +6,6 @@ namespace Content.Server.Mobs;
 /// <summary>
 ///     Mobs with this component will emote a deathgasp when they die.
 /// </summary>
-/// <see cref="DeathgaspSystem"/>
 [RegisterComponent]
 public sealed partial class DeathgaspComponent : Component
 {
