+++ Content.Server/Mobs/Systems/CritMobActionsSystem.cs
@@ -1,4 +1,4 @@
-﻿using Content.Server.Administration;
+using Content.Server.Administration;
 using Content.Server.Chat.Systems;
 using Content.Server.Popups;
 using Content.Server.Speech.Muting;
@@ -9,7 +9,7 @@ using Robust.Server.Console;
 using Robust.Shared.Player;
 using Content.Shared.Speech.Muting;
 
-namespace Content.Server.Mobs;
+namespace Content.Server.Mobs.Systems;
 
 /// <summary>
 ///     Handles performing crit-specific actions.
@@ -17,7 +17,7 @@ namespace Content.Server.Mobs;
 public sealed class CritMobActionsSystem : EntitySystem
 {
     [Dependency] private readonly ChatSystem _chat = default!;
-    [Dependency] private readonly DeathgaspSystem _deathgasp = default!;
+    [Dependency] private readonly MobsSystem _mobsSystem = default!;
     [Dependency] private readonly IServerConsoleHost _host = default!;
     [Dependency] private readonly MobStateSystem _mobState = default!;
     [Dependency] private readonly PopupSystem _popupSystem = default!;
@@ -54,7 +54,7 @@ public sealed class CritMobActionsSystem : EntitySystem
             return;
         }
 
-        args.Handled = _deathgasp.Deathgasp(uid);
+        args.Handled = _mobsSystem.Deathgasp(uid);
     }
 
     private void OnLastWords(EntityUid uid, MobStateActionsComponent component, CritLastWordsEvent args)
