+++ Resources/Locale/en-US/chat/emotes.ftl
@@ -30,6 +30,7 @@ chat-emote-name-monkeyscreeches = Monkey Screeches
 chat-emote-name-robotbeep = Robot
 chat-emote-name-yawn = Yawn
 chat-emote-name-snore = Snore
+chat-emote-name-unconscious = Fall Unconscious
 
 # Message
 chat-emote-msg-scream = screams!
@@ -62,3 +63,4 @@ chat-emote-msg-cathisses = hisses!
 chat-emote-msg-monkeyscreeches = screeches!
 chat-emote-msg-yawn = yawns.
 chat-emote-msg-snore = snores.
+chat-emote-msg-unconscious = staggered, then collapsed, unconscious...
