+++ Resources/Prototypes/Entities/Mobs/base.yml
@@ -97,6 +97,7 @@
       - ActionCritFakeDeath
       - ActionCritLastWords
   - type: Deathgasp
+  - type: Unconscious
   - type: HealthExaminable
     examinableTypes:
     - Blunt
